fix: apply data validity policies in loader

+++ backtesting/data/policy.py
@@ -9,7 +9,7 @@ def expand_monthly_frame(
     if validity != "month_only":
         raise ValueError(f"unsupported validity: {validity}")
 
-    out = pd.DataFrame(index=calendar, columns=frame.columns, dtype="float64")
+    out = pd.DataFrame(index=calendar, columns=frame.columns)
     for ts, row in frame.iterrows():
         month_mask = (calendar.year == ts.year) & (calendar.month == ts.month)
         out.loc[month_mask, :] = row.values
